Add Bishop class and its tests
Added a Bishop class to define the behavior of a bishop,
in this case, We know that without obstacles a bishop can
go to any(with the same color) position in the board
in a max of two moves, to be sure that requested move is
valid was added an arithmetic solution in the validMove method.

In the tests, was added the necessary cases to
ensure that Bishop class is working as expected.